requirements.txt
openai>=1.30.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
mcp>=1.0.0
typer>=0.9.0
rich>=13.0.0
aiohttp>=3.9.0
google-genai>=1.0.0
groq>=0.4.0
# MCP Server dependencies
llama-index==0.12.13
llama-index-embeddings-huggingface
fastmcp==2.14.5
